Unify permission description
