+++ application/frontend/src/pages/users-dashboard/personal-details-box/personal-details-box.tsx
@@ -12,23 +12,9 @@ import {
 } from "@umccr/elsa-constants";
 import { useUiAllowed } from "../../../hooks/ui-allowed";
 import { PermissionDialog } from "../other-users/permission-dialog";
+import { decodeAllowedDescription } from "../users-dashboard-page";
 
 type Props = {};
-
-export function decodeAllowedDescription(allowed: string) {
-  switch (allowed) {
-    case ALLOWED_CHANGE_USER_PERMISSION:
-      return "Allowed to change other user's permission.";
-    case ALLOWED_CREATE_NEW_RELEASE:
-      return "Allowed to create and become a release administrator.";
-    case ALLOWED_OVERALL_ADMIN_VIEW:
-      return "Allowed to view as an app administrator.";
-    case ALLOWED_DATASET_UPDATE:
-      return "Allowed to update/refresh dataset index.";
-    default:
-      return `Unknown 'allowed' code ->${allowed}<-`;
-  }
-}
 
 export const PersonalDetailsBox: React.FC<Props> = ({}) => {
   const [cookies] = useCookies<any>([
